fixed seeds, removed votes, delete comments with posts
workout types added to seed data, delete comments with post data before deleting post, removed vote seeding

+++ controllers/api/post-routes.js
@@ -113,18 +113,27 @@ router.put('/:id', withAuth, (req, res) => {
 });
 
 router.delete('/:id', withAuth, (req, res) => {
-    Post.destroy({ where: { id: req.params.id } })
-        .then(dbPostData => {
-            if (!dbPostData) {
-                res.status(404).json({ message: 'No post found with this id' });
-                return;
-            }
-            res.json(dbPostData);
-        })
+    const commentsDeleted = Comment.destroy({ where: { post_id: req.params.id } })
+        .then(dbPostData => { return true })
         .catch(err => {
-            console.log(err);
+            console.log(err.message);
             res.status(500).json(err);
+            return false;
         });
+    if (commentsDeleted != false) {
+        Post.destroy({ where: { id: req.params.id } })
+            .then(dbPostData => {
+                if (!dbPostData) {
+                    res.status(404).json({ message: 'No post found with this id' });
+                    return;
+                }
+                res.json(dbPostData);
+            })
+            .catch(err => {
+                console.log(err);
+                res.status(500).json(err);
+            });
+    }
 });
 
 module.exports = router;
